moved user transactions to user_routes file
