+++ app/api/user_routes.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, jsonify
 from flask_login import login_required
-from app.models import db, User
+from app.models import db, User, BudgetGroup, BudgetItem, Transaction
 
 user_routes = Blueprint('users', __name__)
 
@@ -17,3 +17,22 @@ def users():
 def user(id):
     user = User.query.get(id)
     return user.to_dict()
+
+# FIND ALL TRANSACTIONS BY USER
+@user_routes.route('/<int:id>/transactions')
+@login_required
+def user_transactions(id):
+    budget_groups = BudgetGroup.query.filter(BudgetGroup.user_id == id).all()
+    budget_group_ids = [budget_group.id for budget_group in budget_groups]
+    transactions_obj = {}
+    for group_id in budget_group_ids:
+        i = BudgetItem.query.get(group_id)
+        if i is not None:
+            i.to_dict()
+            transactions = Transaction.query.filter(
+                Transaction.item_id == i.id).all()
+            if len(transactions) > 0:
+                transactions_obj.update(
+                    {i.id: [transaction.to_dict() for transaction in transactions]})
+
+    return {"transactions": transactions_obj}
